+++ tests/ShapeCrawler.UnitTests/SlideTests.cs
@@ -1,9 +1,13 @@
+using System;
 using System.IO;
 using System.Linq;
+using DocumentFormat.OpenXml.Packaging;
 using FluentAssertions;
+using ShapeCrawler.Enums;
 using ShapeCrawler.Models;
+using ShapeCrawler.Models.Settings;
+using ShapeCrawler.Services.ShapeCreators;
 using Xunit;
-
 
 // ReSharper disable TooManyChainedReferences
 // ReSharper disable TooManyDeclarations
@@ -23,7 +27,7 @@ namespace ShapeCrawler.UnitTests
         public void Hide_HidesSlide()
         {
             // Arrange
-            var pre = PresentationEx.Open(Properties.Resources._001, true);
+            var pre = Presentation.Open(Properties.Resources._001, true);
             var slide = pre.Slides.First();
 
             // Act
@@ -38,7 +42,7 @@ namespace ShapeCrawler.UnitTests
         public void Hidden_ReturnsTrue_WhenSlideIsHidden()
         { 
             // Arrange
-            Slide slide = _fixture.pre002.Slides[2];
+            Slide slide = _fixture.Pre002.Slides[2];
 
             // Act
             bool hidden = slide.Hidden;
@@ -51,7 +55,7 @@ namespace ShapeCrawler.UnitTests
         public async void BackgroundSetImage_ChangesBackground_WhenImageStreamIsPassed()
         {
             // Arrange
-            var pre = new PresentationEx(Properties.Resources._009);
+            var pre = new Presentation(Properties.Resources._009);
             var backgroundImage = pre.Slides[0].Background;
             var imgStream = new MemoryStream(Properties.Resources.test_image_2);
             var bytesBefore = await backgroundImage.GetImageBytesValueTask();
@@ -62,6 +66,45 @@ namespace ShapeCrawler.UnitTests
             // Assert
             var bytesAfter = await backgroundImage.GetImageBytesValueTask();
             bytesAfter.Length.Should().NotBe(bytesBefore.Length);
+        }
+
+        [Fact]
+        public void CustomData_ReturnsData_WhenCustomDataWasAssigned()
+        {
+            // Arrange
+            const string customDataString = "Test custom data";
+            var origPreStream = new MemoryStream();
+            origPreStream.Write(Properties.Resources._001);
+            var originPre = new Presentation(origPreStream, true);
+            var slide = originPre.Slides.First();
+
+            // Act
+            slide.CustomData = customDataString;
+
+            var savedPreStream = new MemoryStream();
+            originPre.SaveAs(savedPreStream);
+            var savedPre = new Presentation(savedPreStream, false);
+            var customData = savedPre.Slides.First().CustomData;
+
+            // Assert
+            customData.Should().Be(customDataString);
+        }
+
+        [Fact]
+        public void Shapes_ReturnsShapeCollectionWithCorrectShapeContentType()
+        {
+            // Arrange
+            var pre = _fixture.Pre003;
+
+            // Act
+            var shapes = pre.Slides.First().Shapes;
+
+            // Assert
+            Assert.Single(shapes.Where(c => c.ContentType.Equals(ShapeContentType.AutoShape)));
+            Assert.Single(shapes.Where(c => c.ContentType.Equals(ShapeContentType.Picture)));
+            Assert.Single(shapes.Where(c => c.ContentType.Equals(ShapeContentType.Table)));
+            Assert.Single(shapes.Where(c => c.ContentType.Equals(ShapeContentType.Chart)));
+            Assert.Single(shapes.Where(c => c.ContentType.Equals(ShapeContentType.Group)));
         }
     }
 }