+++ tests/ShapeCrawler.UnitTests/TestFile_023.cs
@@ -14,7 +14,7 @@ namespace ShapeCrawler.UnitTests
         public void Slide_Shapes_Test()
         {
             // Arrange
-            var pre = new PresentationEx(Properties.Resources._023);
+            var pre = new Presentation(Properties.Resources._023);
 
             // Act-Assert
             var shapes = pre.Slides[0].Shapes;